refactor: token issue 시 인증된 사용자만 가능하게 변경
