+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/access/AccessTokenAuthenticator.java
@@ -1,6 +1,7 @@
 package com.barlow.core.auth.authentication.access;
 
 import com.barlow.core.auth.authentication.core.Authenticator;
+import com.barlow.core.auth.authentication.core.MemberPrincipal;
 import org.springframework.stereotype.Component;
 
 @Component
